+++ canary/apps/react/cra/config-overrides.js
@@ -1,3 +1,0 @@
-const { removeModuleScopePlugin } = require('customize-cra');
-
-module.exports = removeModuleScopePlugin();